Update JetView.ts
Lets assume we are in path /app/page and it gives error when we call setParam("id", 1, true)

Because there is no index in currentUrl array. 

If we create this index when there is no, it allows us to use setParam when we are in path /app/page otherwise we should be in /app/page/xx to use setParam.

I am not an expert on JS. Sorry if this is a bad solution.