[fix] view.refresh doesn't work for views with sub-elements
There were 2 major issues in the previous solution

a)
rendering url was not calculated for top level views
it was expected that for top level views user will call app.refresh
but real usage is different

b)
The subviews were not destroyed but recreated during refresh.
As result old subviews were keeped in memory and complex side-effects were possible.

As part of update app.refresh logic changed as well, not it triggers refresh
of the top view, instead of using its own custom logic.